chore(blink): make new folder for BlinkImpl
